fix: loading databases
It was a bug that if a network has no mailList (or userList), it would carry the lists of the last server.

- created kernel.getDatabases()
- rejects refactored to return when needed

I've forget that Promise.reject() don't end the execution flow (that's why throws works better), so the reject needs to be at the right place, or there is a need to control the execution of rejects in if/else, or you need to force to stop the control flow with a return either after the reject line or returning the reject function.